requirements.txt
Cython==3.1.2
aiogram==3.21.0
asyncpg==0.30.0
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.5
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.28.0
ipywidgets==8.1.7
Jinja2==3.1.6
#jnius==1.1.0
keyring==25.6.0
langchain==0.3.27
langchain_chroma==0.2.5
langchain-community==0.3.27
langchain_core==0.3.72
langchain_huggingface==0.3.1
langchain_text_splitters==0.3.9
langgraph==0.6.2
lockfile==0.12.2
mock==5.2.0
Pillow==11.3.0
protobuf==6.31.1
psycopg2==2.9.10
pyOpenSSL==25.1.0
python-dotenv==1.1.1
pywin32==310
railroad==0.5.0
sentence-transformers==2.7.0
Sphinx==7.2.6
thread==2.0.5
tornado==6.4.1
trove_classifiers==2025.5.9.12
truststore==0.10.3
urllib3_secure_extra==0.1.0
